encoding-faces-within-images/requirements.txt
opencv-python-headless==4.5.3.56
scikit-image==0.19.2
boto3==1.24.2
Pillow
numpy==1.21.6